fixed (#47)
Can't delete if being referenced

+++ app/services/statuses/status_service.py
@@ -1,5 +1,5 @@
 from app import db
-from app.models import Status
+from app.models import Status, Task
 
 
 class StatusService:
@@ -14,9 +14,19 @@ class StatusService:
         return status
 
     def delete(self, id):
-        Status.query.filter_by(id=id).delete()
-        db.session.commit()
-        return {"success": True}
+        """
+        Check if there are any references to this status
+        """
+        count = Task.query.filter_by(status_id=id).count()
+        if count == 0:
+            Status.query.filter_by(id=id).delete()
+            db.session.commit()
+            return {"success": True}
+        else:
+            return {
+                "success": False,
+                "reason": "This status is being referenced"
+            }
 
     def put(self, id, title):
         status = Status.query.filter_by(id=id).first()
